% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapp_server.R
\name{app server}
\alias{app server}
\alias{web-app-server}
\alias{app-server}
\alias{server}
\alias{app_server}
\alias{initial_reactive_values}
\alias{initial_output}
\title{Generate the Server Code for the Web App}
\usage{
app_server(input, output, session)

initial_reactive_values()

initial_output(rv, output)
}
\arguments{
\item{input}{\code{input} \code{list} for the UI.}

\item{output}{\code{output} \code{list} for the UI.}

\item{session}{Environment for the UI.}

\item{rv}{\code{\link[shiny:reactiveValues]{reactiveValues}} \code{list} for the UI.}
}
\value{
\code{app_server}: an observer reference class object (see \code{\link[shiny:observeEvent]{observeEvent}} and \code{\link[shiny:observe]{observe}}). \cr
\code{initial_reactive_values}: a \code{\link[shiny:reactiveValues]{reactiveValues}} \code{list}. \cr
\code{initial_output}: an \code{output} \code{list}. \cr
}
\description{
Generate the Server Code for the Web App
}
\seealso{
Other shinyapp: 
\code{\link{app ui}},
\code{\link{app}}
}
\concept{shinyapp}
