% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapp_ui.R
\name{app ui}
\alias{app ui}
\alias{web-app-ui}
\alias{app-ui}
\alias{ui}
\alias{app_ui}
\alias{app_theme}
\title{Generate the User Interface for the Web App}
\usage{
app_ui(global = global_list(main = "."))

app_theme()
}
\arguments{
\item{global}{A \code{list} of global values for the app.}
}
\value{
A UI definition, component shiny tags, or bootswatch theme.
}
\description{
\code{app_ui} constructs the user interface (UI) for the web application
See \code{Details} for hierarchy of functions.
}
\details{
The UI is hierarchical built as:
\itemize{
\item \code{app_ui}
}
}
\seealso{
Other shinyapp: 
\code{\link{app server}},
\code{\link{app}}
}
\concept{shinyapp}
