% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapp.R
\name{app}
\alias{app}
\alias{web-app}
\alias{run_app}
\alias{deploy_app}
\alias{global_list}
\title{Build and Launch the Web Application}
\usage{
run_app(main = ".")

deploy_app(main = ".")

global_list(main = ".")
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}
}
\value{
\code{global_list}: a \code{list} of values that will be globally available in the app. \cr
\code{run_app} runs the app itself.
}
\description{
\code{run_app} constructs and launches a local version of the web application by running \code{\link[shiny:runApp]{shiny::runApp}} pointed to the \code{app} subdirectory.
\code{global_list} creates a list of values that are globally available in an app run.
}
\seealso{
Other shinyapp: 
\code{\link{app server}},
\code{\link{app ui}}
}
\concept{shinyapp}
