% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker-core.R
\name{check_app_directory}
\alias{check_app_directory}
\title{Check if a directory exists and is valid}
\usage{
check_app_directory(app_dir)
}
\arguments{
\item{app_dir}{Character. Path to check.}
}
\value{
Invisibly returns TRUE if valid, otherwise aborts with an error message.
}
\description{
Check if a directory exists and is valid
}
\keyword{internal}
