% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{create_dockerignore}
\alias{create_dockerignore}
\title{Create a .dockerignore file}
\usage{
create_dockerignore(output_dir)
}
\arguments{
\item{output_dir}{Character. Directory where the file should be created.}
}
\value{
Character. Path to the created file.
}
\description{
Create a .dockerignore file
}
\keyword{internal}
