% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-detection.R
\name{detect_app_type}
\alias{detect_app_type}
\title{Detect the type of Shiny application (R or Python)}
\usage{
detect_app_type(app_dir)
}
\arguments{
\item{app_dir}{Character. Path to the Shiny application directory.}
}
\value{
Character. Either "r" or "python".
}
\description{
Detect the type of Shiny application (R or Python)
}
\keyword{internal}
