% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-detection.R
\name{detect_dependencies}
\alias{detect_dependencies}
\title{Detect dependencies for a Shiny application}
\usage{
detect_dependencies(app_dir, app_type)
}
\arguments{
\item{app_dir}{Character. Path to the Shiny application directory.}

\item{app_type}{Character. Either "r" or "python".}
}
\value{
Character vector with detected dependencies.
}
\description{
Detect dependencies for a Shiny application
}
\keyword{internal}
