% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-detection.R
\name{detect_python_dependencies}
\alias{detect_python_dependencies}
\title{Detect Python package dependencies}
\usage{
detect_python_dependencies(app_dir)
}
\arguments{
\item{app_dir}{Character. Path to the Shiny application directory.}
}
\value{
Character vector of detected Python package dependencies.
}
\description{
Detect Python package dependencies
}
\keyword{internal}
