% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-detection.R
\name{detect_r_dependencies}
\alias{detect_r_dependencies}
\title{Detect R package dependencies}
\usage{
detect_r_dependencies(app_dir)
}
\arguments{
\item{app_dir}{Character. Path to the Shiny application directory.}
}
\value{
Character vector of detected R package dependencies.
}
\description{
Detect R package dependencies
}
\keyword{internal}
