% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{directory creation}
\alias{directory creation}
\alias{create}
\alias{setup}
\alias{update}
\alias{directory}
\alias{directory-setup}
\alias{directory-creation}
\alias{directory-update}
\alias{create_dir}
\alias{update_dir}
\alias{setup_dir}
\title{Create or Update the Structure of a Directory and Fill It with Content or Update an Existing Directory}
\usage{
create_dir(
  main = ".",
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

update_dir(
  main = ".",
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

setup_dir(
  main = ".",
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree. Default value (\code{"."}) roots the directory in the present location.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}}.}

\item{quiet}{\code{logical} indicator if progress messages should be quieted.}

\item{verbose}{\code{logical} indicator of whether or not to produce all of the messages.}
}
\value{
The \code{list} of directory settings \code{\link[base:invisible]{invisible}}-ly.
}
\description{
Instantiates the necessary folder structure for a directory, writes the setup configuration file, and fills the directory with content. \cr
Options for pre-defined setups include \code{setup_sandbox} for quick and flexible builds and \code{setup_production} for robust, rigid builds, as defined in \code{\link{directory_settings}}. \cr
\code{update_dir} updates an existing directory and \code{update_sandbox} and \code{update_production} are companions akin to their \verb{setup_<>} functions.
}
\seealso{
Other orchestration: 
\code{\link{directory configuration file}},
\code{\link{directory paths}},
\code{\link{directory settings}}
}
\concept{orchestration}
