% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{directory filling}
\alias{directory filling}
\alias{fill}
\alias{fill_dir}
\alias{fill_app}
\title{Fill a Portalcasting Directory with Basic Components}
\usage{
fill_dir(main = ".")

fill_app(main = ".")
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}
}
\value{
\code{NULL}, \code{\link[base:invisible]{invisible}}-ly.
}
\description{
Fill the directory with components including:
* Web Application (\code{\link{fill_app}})
* transfers app files from package to main
}
\seealso{
Other orchestration: 
\code{\link{directory configuration file}},
\code{\link{directory creation}},
\code{\link{directory paths}},
\code{\link{directory settings}}
}
\concept{content-prep}
\concept{orchestration}
