% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{directory paths}
\alias{directory paths}
\alias{paths}
\alias{app_path}
\alias{directory_configuration_path}
\title{Create a List of Full Directory Paths}
\usage{
paths(main = ".")

app_path(main = ".")

directory_configuration_path(main = ".")
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}
}
\value{
\code{list} of directory paths or specific \code{character} paths.
}
\description{
Upon creation (or updating) of the directory, all the standard file and subdirectory paths are set based on \code{\link{directory_settings}}. \cr
\code{paths} produces the full path \code{list}, whose contents can then also be accessed with specialized functions, see \code{Details}.
}
\details{
Wrapper functions for specific subdirectories and files include:
\itemize{
\item Files
\itemize{
\item \code{directory_configuration_path}
}
\item Subdirectories
\itemize{
\item \code{app_path}
}
}
}
\seealso{
Other orchestration: 
\code{\link{directory configuration file}},
\code{\link{directory creation}},
\code{\link{directory filling}},
\code{\link{directory settings}}
}
\concept{orchestration}
