% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{directory settings}
\alias{directory settings}
\alias{settings}
\alias{directory_settings}
\alias{directory_files}
\alias{directory_subdirectories}
\title{Create a Directory Settings List}
\usage{
directory_settings(
  files = directory_files(),
  subdirectories = directory_subdirectories(),
  save = TRUE,
  overwrite = TRUE,
  force = FALSE
)

directory_files()

directory_subdirectories(app = "app")
}
\arguments{
\item{files}{\code{list} of \code{character} names of standard files, see \code{\link{directory_files}}.}

\item{subdirectories}{\code{list} of \code{character} names of standard subdirectories, see \code{\link{directory_subdirectories}}.}

\item{save}{\code{logical} indicator controlling if the output should be saved out.}

\item{overwrite}{\code{logical} indicator of whether or not file writing should occur even if a local copy already exists.}

\item{force}{\code{logical} indicator of whether or not existing files or folders (such as the archive) should be over-written if an up-to-date copy exists (most users should leave as \code{FALSE}).}

\item{app}{\code{character} name for the app subdirectory.}
}
\value{
Named \code{list} of settings for the directory (for \code{directory_settings}) or \code{list} of settings components (for \code{directory_files}, \code{directory_subdirectories}).
}
\description{
Compose parts of and complete directory settings controls lists. \cr \cr
Most users will not want or need to change the directory folders and file names, but it is helpful to have them be flexible for certain circumstances, and \code{directory_files}, \code{directory_subdirectories}, and \code{directory_settings} gather them into a set of lists for pipeline functionality. \cr
}
\examples{
   directory_settings( )
   directory_files( )
   directory_subdirectories( )


}
\seealso{
Other orchestration: 
\code{\link{directory configuration file}},
\code{\link{directory creation}},
\code{\link{directory paths}}
}
\concept{customization}
\concept{orchestration}
