% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-detection.R
\name{extract_namespace_packages}
\alias{extract_namespace_packages}
\title{Extract package names from namespace calls (package::function)}
\usage{
extract_namespace_packages(r_code)
}
\arguments{
\item{r_code}{Character vector. R code to analyze.}
}
\value{
Character vector of package names.
}
\description{
Extract package names from namespace calls (package::function)
}
\keyword{internal}
