% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_available_port}
\alias{find_available_port}
\title{Find an available port starting from a given port}
\usage{
find_available_port(start_port, max_tries = 100)
}
\arguments{
\item{start_port}{Integer. The port to start checking from.}

\item{max_tries}{Integer. Maximum number of ports to check. Default: 100.}
}
\value{
Integer. An available port or NULL if no port found.
}
\description{
Find an available port starting from a given port
}
\keyword{internal}
