% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{find_containers_for_app}
\alias{find_containers_for_app}
\title{Find containers associated with an application}
\usage{
find_containers_for_app(app_dir, quiet = FALSE, ...)
}
\arguments{
\item{app_dir}{Character. Path to the application directory.}

\item{quiet}{Logical. Whether to suppress messages.}

\item{...}{Additional arguments passed to processx.}
}
\value{
Character vector. Container IDs.
}
\description{
Find containers associated with an application
}
\keyword{internal}
