% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{format_dependencies}
\alias{format_dependencies}
\title{Format dependencies for inclusion in Dockerfile}
\usage{
format_dependencies(app_type, dependencies)
}
\arguments{
\item{app_type}{Character. Either "r" or "python".}

\item{dependencies}{List of dependencies.}
}
\value{
Character string with formatted dependencies.
}
\description{
Format dependencies for inclusion in Dockerfile
}
\keyword{internal}
