% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{format_env_vars}
\alias{format_env_vars}
\title{Format environment variables for inclusion in Dockerfile}
\usage{
format_env_vars(env_vars)
}
\arguments{
\item{env_vars}{Named character vector. Environment variables.}
}
\value{
Character string with formatted environment variables.
}
\description{
Format environment variables for inclusion in Dockerfile
}
\keyword{internal}
