% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_app_url}
\alias{generate_app_url}
\title{Generate app URL based on app type and port}
\usage{
generate_app_url(app_type, port)
}
\arguments{
\item{app_type}{Character. Either "r" or "python".}

\item{port}{Integer. Port number.}
}
\value{
Character. URL to access the app.
}
\description{
Generate app URL based on app type and port
}
\keyword{internal}
