% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{generate_image_tag}
\alias{generate_image_tag}
\title{Generate a default Docker image tag based on app directory}
\usage{
generate_image_tag(app_dir)
}
\arguments{
\item{app_dir}{Character. Path to the Shiny application directory.}
}
\value{
Character. Generated tag.
}
\description{
Generate a default Docker image tag based on app directory
}
\keyword{internal}
