% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R
\name{get_app_issues}
\alias{get_app_issues}
\title{Get issues and recommendations for Shiny app containerization}
\usage{
get_app_issues(results)
}
\arguments{
\item{results}{List. Results from Shiny app checks}
}
\value{
List with issues and recommendations
}
\description{
Get issues and recommendations for Shiny app containerization
}
\keyword{internal}
