% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R
\name{get_docker_issues}
\alias{get_docker_issues}
\title{Get Docker environment issues and recommendations}
\usage{
get_docker_issues(results)
}
\arguments{
\item{results}{List. Results from Docker environment checks}
}
\value{
List with issues and recommendations
}
\description{
Get Docker environment issues and recommendations
}
\keyword{internal}
