% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker-core.R
\name{get_platform_flag}
\alias{get_platform_flag}
\title{Get platform flag for Docker if needed}
\usage{
get_platform_flag(app_type)
}
\arguments{
\item{app_type}{Character. Either "r" or "python".}
}
\value{
Character vector. Platform flag for Docker commands or empty vector if not needed.
}
\description{
Returns platform flag for Docker commands if running on ARM64 and if app type is R.
This is needed because Shiny Server isn't available natively for ARM64 yet.
}
\keyword{internal}
