% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{handle_multiple_containers}
\alias{handle_multiple_containers}
\title{Handle stopping multiple containers}
\usage{
handle_multiple_containers(container_ids, quiet = FALSE, force = FALSE, ...)
}
\arguments{
\item{container_ids}{Character vector. Container IDs.}

\item{quiet}{Logical. Whether to suppress output.}

\item{force}{Logical. Whether to skip confirmation.}

\item{...}{Additional arguments passed to processx.}
}
\value{
Logical. TRUE if successful.
}
\description{
Handle stopping multiple containers
}
\keyword{internal}
