% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_port_configuration}
\alias{handle_port_configuration}
\title{Handle port configuration and availability}
\usage{
handle_port_configuration(port, force_port = FALSE)
}
\arguments{
\item{port}{Integer. Requested port.}

\item{force_port}{Logical. Whether to enforce the requested port.}
}
\value{
Integer. The port to use, either the requested one or an alternative.
}
\description{
Handle port configuration and availability
}
\keyword{internal}
