% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R
\name{health_dependencies}
\alias{health_dependencies}
\title{Report dependencies health in a Shiny app}
\usage{
health_dependencies(app_dir, app_type, verbose = FALSE)
}
\arguments{
\item{app_dir}{Character. Path to the Shiny application directory}

\item{app_type}{Character. Either "r" or "python"}

\item{verbose}{Logical. If TRUE, provide more detailed output}
}
\value{
List with dependency details
}
\description{
Report dependencies health in a Shiny app
}
\keyword{internal}
