% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R
\name{health_docker_compose}
\alias{health_docker_compose}
\title{Report Docker Compose availability health}
\usage{
health_docker_compose(timeout = 30, ...)
}
\arguments{
\item{timeout}{Numeric. Timeout in seconds for commands}

\item{...}{Additional arguments passed to processx}
}
\value{
List with Docker Compose details
}
\description{
Report Docker Compose availability health
}
\keyword{internal}
