% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R
\name{health_docker_daemon}
\alias{health_docker_daemon}
\title{Report Docker daemon health status}
\usage{
health_docker_daemon(timeout = 30, ...)
}
\arguments{
\item{timeout}{Numeric. Timeout in seconds for commands}

\item{...}{Additional arguments passed to processx}
}
\value{
List with Docker daemon status details
}
\description{
Report Docker daemon health status
}
\keyword{internal}
