% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R
\name{health_docker_installation}
\alias{health_docker_installation}
\title{Report Docker installation health}
\usage{
health_docker_installation(timeout = 30, ...)
}
\arguments{
\item{timeout}{Numeric. Timeout in seconds for commands}

\item{...}{Additional arguments passed to processx}
}
\value{
List with Docker installation details
}
\description{
Report Docker installation health
}
\keyword{internal}
