% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R
\name{health_docker_permissions}
\alias{health_docker_permissions}
\title{Report Docker permissions health}
\usage{
health_docker_permissions(timeout = 30, ...)
}
\arguments{
\item{timeout}{Numeric. Timeout in seconds for commands}

\item{...}{Additional arguments passed to processx}
}
\value{
List with Docker permission details
}
\description{
Report Docker permissions health
}
\keyword{internal}
