% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R
\name{health_port_availability}
\alias{health_port_availability}
\title{Report port availability health}
\usage{
health_port_availability(ports, timeout = 30, port_timeout = 2, ...)
}
\arguments{
\item{ports}{Integer vector. Ports to check}

\item{timeout}{Numeric. Timeout in seconds for commands}

\item{port_timeout}{Numeric. Timeout in seconds for port checking}

\item{...}{Additional arguments passed to processx}
}
\value{
List with port availability details
}
\description{
Report port availability health
}
\keyword{internal}
