% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R
\name{health_shiny_app}
\alias{health_shiny_app}
\title{Report Shiny app structure health}
\usage{
health_shiny_app(app_dir, verbose = FALSE)
}
\arguments{
\item{app_dir}{Character. Path to the Shiny application directory}

\item{verbose}{Logical. If TRUE, provide more detailed output}
}
\value{
List with Shiny app details
}
\description{
Report Shiny app structure health
}
\keyword{internal}
