% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R
\name{health_system_info}
\alias{health_system_info}
\title{Report system information health}
\usage{
health_system_info()
}
\value{
List with system information details
}
\description{
Report system information health
}
\keyword{internal}
