% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker-core.R
\name{is_arm64}
\alias{is_arm64}
\title{Detect if system is running on ARM64 architecture}
\usage{
is_arm64()
}
\value{
Logical. TRUE if the system is running on ARM64 architecture.
}
\description{
Detect if system is running on ARM64 architecture
}
\keyword{internal}
