% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_port_available}
\alias{is_port_available}
\title{Check if a port is available on localhost}
\usage{
is_port_available(port)
}
\arguments{
\item{port}{Integer. The port to check.}
}
\value{
Logical. TRUE if the port is available, FALSE if already in use.
}
\description{
Check if a port is available on localhost
}
\keyword{internal}
