% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{prepare_build_arguments}
\alias{prepare_build_arguments}
\title{Prepare Docker build arguments}
\usage{
prepare_build_arguments(app_dir, tag, build_args = NULL, quiet = FALSE)
}
\arguments{
\item{app_dir}{Character. Path to application directory.}

\item{tag}{Character. Image tag.}

\item{build_args}{Named character vector. Additional build arguments.}

\item{quiet}{Logical. Whether to use quiet mode.}
}
\value{
Character vector of build arguments.
}
\description{
Prepare Docker build arguments
}
\keyword{internal}
