% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{process_containers}
\alias{process_containers}
\title{Process and remove containers}
\usage{
process_containers(container_lines, include_running, force, quiet, ...)
}
\arguments{
\item{container_lines}{Character vector. Container data lines.}

\item{include_running}{Logical. Whether to include running containers.}

\item{force}{Logical. Whether to force removal.}

\item{quiet}{Logical. Whether to suppress output.}

\item{...}{Additional arguments passed to processx.}
}
\value{
Integer. Number of removed containers.
}
\description{
Process and remove containers
}
\keyword{internal}
