% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{remove_shinydocker_images}
\alias{remove_shinydocker_images}
\title{Remove shinydocker images}
\usage{
remove_shinydocker_images(app_dir, force, quiet, ...)
}
\arguments{
\item{app_dir}{Character. Path to the application directory. If NULL, all shinydocker images.}

\item{force}{Logical. Whether to force removal.}

\item{quiet}{Logical. Whether to suppress output.}

\item{...}{Additional arguments passed to processx.}
}
\value{
Integer. Number of removed images.
}
\description{
Remove shinydocker images
}
\keyword{internal}
