% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{run_docker_prune}
\alias{run_docker_prune}
\title{Run docker system prune}
\usage{
run_docker_prune(force, quiet, ...)
}
\arguments{
\item{force}{Logical. Whether to force prune.}

\item{quiet}{Logical. Whether to suppress output.}

\item{...}{Additional arguments passed to processx.}
}
\value{
Invisibly returns TRUE.
}
\description{
Run docker system prune
}
\keyword{internal}
