% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{run_with_compose}
\alias{run_with_compose}
\title{Run container using docker-compose}
\usage{
run_with_compose(
  app_dir,
  port,
  env_vars = NULL,
  detach = FALSE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{app_dir}{Character. Path to the Shiny application directory.}

\item{port}{Integer. The port to map.}

\item{env_vars}{Named character vector. Environment variables.}

\item{detach}{Logical. Whether to run in detached mode.}

\item{quiet}{Logical. Whether to suppress output.}

\item{...}{Additional arguments passed to processx.}
}
\value{
Character. Container ID if successful.
}
\description{
Run container using docker-compose
}
\keyword{internal}
