% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDocker.R
\docType{package}
\name{ShinyDocker-package}
\alias{ShinyDocker}
\alias{ShinyDocker-package}
\title{ShinyDocker: R Package Shiny App in Docker}
\description{
Toy example package with a shiny app that is dockerized and served.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dapperstats.github.io/ShinyDocker/}
  \item \url{https://github.com/dapperstats/ShinyDocker}
  \item Report bugs at \url{https://github.com/dapperstats/ShinyDocker/issues}
}

}
\author{
\strong{Maintainer}: Juniper L. Simonis \email{simonis@dapperstats.com} (\href{https://orcid.org/0000-0001-9798-0460}{ORCID})

}
\keyword{internal}
