% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{should_use_compose}
\alias{should_use_compose}
\title{Determine if docker-compose should be used for container operations}
\usage{
should_use_compose(app_dir, docker_compose = TRUE)
}
\arguments{
\item{app_dir}{Character. Path to the Shiny application directory.}

\item{docker_compose}{Logical. User preference for using docker-compose.}
}
\value{
Logical. TRUE if docker-compose should be used.
}
\description{
Determine if docker-compose should be used for container operations
}
\keyword{internal}
