% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{stop_all_containers}
\alias{stop_all_containers}
\title{Stop all running Docker containers}
\usage{
stop_all_containers(quiet = FALSE, force = FALSE, ...)
}
\arguments{
\item{quiet}{Logical. Whether to suppress output.}

\item{force}{Logical. Whether to skip confirmation.}

\item{...}{Additional arguments passed to processx.}
}
\value{
Logical. TRUE if successful.
}
\description{
Stop all running Docker containers
}
\keyword{internal}
