% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{stop_docker_container}
\alias{stop_docker_container}
\title{Stop a Docker container}
\usage{
stop_docker_container(container_id, quiet = FALSE, ...)
}
\arguments{
\item{container_id}{Character. ID of the container to stop.}

\item{quiet}{Logical. Whether to suppress output.}

\item{...}{Additional arguments passed to processx.}
}
\value{
Logical. TRUE if successful.
}
\description{
Stop a Docker container
}
\keyword{internal}
