% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{stop_running_containers}
\alias{stop_running_containers}
\title{Stop running containers}
\usage{
stop_running_containers(container_ids, quiet, ...)
}
\arguments{
\item{container_ids}{Character vector. Container IDs to stop.}

\item{quiet}{Logical. Whether to suppress output.}

\item{...}{Additional arguments passed to processx.}
}
\value{
Invisibly returns TRUE.
}
\description{
Stop running containers
}
\keyword{internal}
