% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{stop_with_compose}
\alias{stop_with_compose}
\title{Stop containers using docker-compose}
\usage{
stop_with_compose(app_dir, quiet = FALSE, ...)
}
\arguments{
\item{app_dir}{Character. Path to the application directory.}

\item{quiet}{Logical. Whether to suppress output.}

\item{...}{Additional arguments passed to processx.}
}
\value{
Logical. TRUE if successful.
}
\description{
Stop containers using docker-compose
}
\keyword{internal}
