% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{update_compose_port}
\alias{update_compose_port}
\title{Update port mapping in docker-compose.yml}
\usage{
update_compose_port(app_dir, port)
}
\arguments{
\item{app_dir}{Character. Path to the application directory.}

\item{port}{Integer. New port to use.}
}
\value{
Invisibly returns TRUE if successful.
}
\description{
Update port mapping in docker-compose.yml
}
\keyword{internal}
