% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapp_server.R
\name{app server}
\alias{app server}
\alias{web-app-server}
\alias{app-server}
\alias{server}
\alias{app_server}
\alias{initial_reactive_values}
\alias{initial_output}
\title{Generate the Server Code for the Web App}
\usage{
app_server(input, output, session)

initial_reactive_values(global = global_list())

initial_output(main = ".", global, rv, output)
}
\arguments{
\item{input}{\code{input} \code{list} for the UI.}

\item{output}{\code{output} \code{list} for the UI.}

\item{session}{Environment for the UI.}

\item{global}{A \code{list} of global values for the app.}

\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{rv}{\code{\link[shiny:reactiveValues]{reactiveValues}} \code{list} for the UI.}

\item{event}{\code{character} value of the server event. \cr}
}
\value{
\code{app_server}: an observer reference class object (see \code{\link[shiny:observeEvent]{observeEvent}} and \code{\link[shiny:observe]{observe}}). \cr
\code{initial_reactive_values}: a \code{\link[shiny:reactiveValues]{reactiveValues}} \code{list}. \cr
\code{initial_output}: an \code{output} \code{list}. \cr
}
\description{
Generate the Server Code for the Web App
}
\examples{
\dontrun{
   main1 <- file.path(tempdir(), "app")
   setup_dir(main = main1)

   global <- global_list(main = main1)
 
   rv     <- initial_reavtive_values(global = global)

   output <- initial_output(main   = main1,
                            global = global,
                            rv     = rv,
                            output = list())

   unlink(main1, recursive = TRUE)
 }

}
\seealso{
Other shinyapp: 
\code{\link{app ui}},
\code{\link{app}}
}
\concept{shinyapp}
